requirements.txt
aiohttp[speedups]~=3.7.4

requests~=2.25.1
PyYAML~=5.4.1